Project_feedback_nathan
This project is quite thorough!  Your intro is well done.  The data checks out.  One figure is a bit off, the State names are all scrunched together.  It's your first figure, the census tract/eligibility one.  Other than that, they all look great.
The website built properly, although you had to download an R package from github, because it is not hosted on CRAN it was a teachable moment for Dr. Wilson.